requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple

torch
torchvision
torchaudio

inference-sdk
tqdm
pillow
pandas
opencv-python
matplotlib
requests
editdistance
scikit-learn


fastapi
uvicorn
python-multipart

